♻️ refactor: isSelectedList 상태 끌어올리기
